requirements.txt
arrow==0.17.0
asgiref==3.3.1
blessed==1.17.12
boto3==1.16.37
botocore==1.19.37
Django==3.1.4
django-picklefield==3.0.1
django-q==1.3.4
django-storages==1.10.1
form==0.0.1
generics==3.0.0
jmespath==0.10.0
Pillow==8.0.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wcwidth==0.2.5
